+++ test/helpers.dart
@@ -8,5 +8,8 @@ final successPredicate = predicate<SecretaryEvent>((e) => e.isSuccess);
 Matcher hasKey<K, T>(K key) =>
     predicate<SecretaryEvent<K, T>>((e) => e.key == key);
 
-Future<T> delayedValue<T>(T value, Duration delay) =>
+Future<T> delayedValue<T>(
+  T value, [
+  Duration delay = const Duration(milliseconds: 500),
+]) =>
     Future.delayed(delay, () => value);